Refactor key utils out of nbhtmlcache
